refactor(.zshrc): use `hash` to check if nvim is installed
`hash` is a built in bash command that contains all the command names
and paths. It act like a cache fo commands, whenever we write a command
the shell checks if it is in the hash table, if not search in the PATH
and if found adds it to the has table.