feat(.zsh-tmux): add dnd to tmux config, enhanced alredy dnd alias to function
The dnd alias has become a function, if passed a directory path it will
list all the content if passed a file path it will list only that file.

The same function has been added to tmux configuration. While in tmux
you can double click on a file and it will open a small window with it
inside, this enables you to drag and drop.